Add a Suffix to Variables in a List of Expressions

Add a suffix to variables in a list of expressions

Usage

add_suffix_to_vars(order, vars, suffix)

Arguments

order

List of expressions

Permitted Values: list of variables or desc(<variable>) function calls created by exprs(), e.g., exprs(ADT, desc(AVAL))

vars

Variables to change

Permitted Values: list of variables created by exprs()

suffix

Suffix

Permitted Values: A character scalar

Value

The list of expression where for each element the suffix (suffix) is added to every symbol specified for vars

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(rlang)

add_suffix_to_vars(exprs(ADT, desc(AVAL), AVALC), vars = exprs(AVAL), suffix = ".join")
#> [[1]]
#> ADT
#> 
#> [[2]]
#> desc(AVAL.join)
#> 
#> [[3]]
#> AVALC
#>