Is an Argument a Filter Condition?

Is an Argument a Filter Condition?

assert_filter_cond(arg, optional = FALSE)

Arguments

arg

Quosure - filtering condition.

optional

Logical - is the argument optional? Defaults to FALSE.

Value

Performs necessary checks and returns arg if all pass. Otherwise throws an informative error.

Details

Check if arg is a suitable filtering condition to be used in functions like subset or dplyr::filter.

See also

Author

Ondrej Slama

Examples

library(admiral.test)
data(admiral_dm)

# typical usage in a function as a parameter check
example_fun <- function(dat, x) {
  x <- assert_filter_cond(rlang::enquo(x))
  dplyr::filter(dat, !!x)
}

example_fun(admiral_dm, AGE == 64)
#> # A tibble: 5 x 25
#>   STUDYID DOMAIN USUBJID SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>   <chr>   <chr>  <chr>   <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#> 1 CDISCP… DM     01-701… 1023   2012-0… 2012-0… 2012-08… 2012-09… ""     
#> 2 CDISCP… DM     01-703… 1042   2013-0… 2013-0… 2013-03… 2013-08… ""     
#> 3 CDISCP… DM     01-706… 1041   2013-1… 2014-0… 2013-12… 2014-07… ""     
#> 4 CDISCP… DM     01-713… 1179   2013-0… 2014-0… 2013-08… 2014-02… ""     
#> 5 CDISCP… DM     01-717… 1344   2014-0… 2014-0… 2014-01… 2014-03… ""     
#> # … with 16 more variables: RFPENDTC <chr>, DTHDTC <chr>, DTHFL <chr>,
#> #   SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>, ETHNIC <chr>,
#> #   ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>,
#> #   DMDTC <chr>, DMDY <dbl>

try(example_fun(admiral_dm, USUBJID))
#> Error in assert_filter_cond(rlang::enquo(x)) : 
#>   `x` must be a filter condition but is a symbol