Is an Argument a Logical Scalar (Boolean)?

Checks if an argument is a logical scalar

Usage

assert_logical_scalar(
  arg,
  optional = FALSE,
  arg_name = rlang::caller_arg(arg),
  message = NULL,
  class = "assert_logical_scalar",
  call = parent.frame()
)

Arguments

arg

A function argument to be checked

Default value

none

optional

Is the checked argument optional?
If set to FALSE and arg is NULL then an error is thrown. Otherwise, NULL is considered as valid value.

Default value

FALSE

arg_name

string indicating the label/symbol of the object being checked.

Default value

rlang::caller_arg(arg)

message

string passed to cli::cli_abort(message). When NULL, default messaging is used (see examples for default messages). "{arg_name}" can be used in messaging.

Default value

NULL

class

Subclass of the condition.

call

The execution environment of a currently running function, e.g. call = caller_env(). The corresponding function call is retrieved and mentioned in error messages as the source of the error.

You only need to supply call when throwing a condition from a helper function which wouldn't be relevant to mention in the message.

Can also be NULL or a defused function call to respectively not display any call or hard-code a code to display.

For more information about error calls, see Including function calls in error messages.

Value

The function throws an error if arg is neither TRUE or FALSE. Otherwise, the input is returned invisibly.

See also

Examples

example_fun <- function(flag) {
  assert_logical_scalar(flag)
}

example_fun(FALSE)

try(example_fun(NA))
#> Error in example_fun(NA) : 
#>   Argument `flag` must be either TRUE or FALSE, but is `NA`.

try(example_fun(c(TRUE, FALSE, FALSE)))
#> Error in example_fun(c(TRUE, FALSE, FALSE)) : 
#>   Argument `flag` must be either TRUE or FALSE, but is a logical vector.

try(example_fun(1:10))
#> Error in example_fun(1:10) : 
#>   Argument `flag` must be either TRUE or FALSE, but is an integer vector.