Is an Argument a List of Variables?

Checks if an argument is a valid list of variables created using vars()

assert_vars(arg, optional = FALSE)

Arguments

arg

A function argument to be checked

optional

Is the checked parameter optional? If set to FALSE and arg is NULL then an error is thrown

Value

The function throws an error if arg is not a list of variables created using vars()and returns the input invisibly otherwise.

See also

Author

Samia Kabi

Examples

example_fun <- function(by_vars) {
  assert_vars(by_vars)
}

example_fun(vars(USUBJID, PARAMCD))

try(example_fun(rlang::exprs(USUBJID, PARAMCD)))
#> Error in assert_vars(by_vars) : 
#>   `by_vars` must be a list of unquoted variable names, e.g. `vars(USUBJID, VISIT)`

try(example_fun(c("USUBJID", "PARAMCD", "VISIT")))
#> Error in assert_vars(by_vars) : 
#>   `by_vars` must be a list of unquoted variable names, e.g. `vars(USUBJID, VISIT)`

try(example_fun(vars(USUBJID, toupper(PARAMCD), desc(AVAL))))
#> Error in assert_vars(by_vars) : 
#>   `by_vars` must be a list of unquoted variable names, e.g. `vars(USUBJID, VISIT)`, but the following elements are not: `toupper(PARAMCD)` and `desc(AVAL)`