Get Constant Variables

Get Constant Variables

get_constant_vars(dataset, by_vars, ignore_vars = NULL)

Arguments

dataset

A data frame.

by_vars

By variables The groups defined by the by variables are considered separately. I.e., if a variable is constant within each by group, it is returned.

ignore_vars

Variables to ignore The specified variables are not considered, i.e., they are not returned even if they are constant (unless they are included in the by variables).

Permitted Values: A list of variable names or selector function calls like starts_with("EX")

Value

Variable vector.

See also

Examples

library(admiral.test)
data(admiral_vs)

admiral:::get_constant_vars(admiral_vs, by_vars = vars(USUBJID, VSTESTCD))
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^USUBJID
#> env:  0x55f6e3b399e8
#> 
#> [[2]]
#> <quosure>
#> expr: ^VSTESTCD
#> env:  0x55f6e3b399e8
#> 
#> [[3]]
#> <quosure>
#> expr: ^STUDYID
#> env:  0x55f6e3c70d58
#> 
#> [[4]]
#> <quosure>
#> expr: ^DOMAIN
#> env:  0x55f6e3c70d58
#> 
#> [[5]]
#> <quosure>
#> expr: ^VSTEST
#> env:  0x55f6e3c70d58
#> 

admiral:::get_constant_vars(
  admiral_vs,
  by_vars = vars(USUBJID, VSTESTCD),
  ignore_vars = vars(DOMAIN, tidyselect::starts_with("VS"))
)
#> <list_of<quosure>>
#> 
#> [[1]]
#> <quosure>
#> expr: ^USUBJID
#> env:  0x55f6e3b399e8
#> 
#> [[2]]
#> <quosure>
#> expr: ^VSTESTCD
#> env:  0x55f6e3b399e8
#> 
#> [[3]]
#> <quosure>
#> expr: ^STUDYID
#> env:  0x55f6dfe79010
#>