Negated Value Matching

Returns a logical vector indicating if there is no match of the left operand in the right operand.

Usage

x %notin% table

Arguments

x

The values to be matched

table

The values to be matched against

Value

A logical vector

See also