Or

This function is deprecated. Please get in touch if you are using this function!

Usage

lhs %or% rhs

Arguments

lhs

Any valid R expression

Default value

none

rhs

Any valid R expression

Default value

none

Value

Either the result of evaluating lhs, rhs or an error

Details

The function evaluates the expression lhs and if this expression results in an error, it catches that error and proceeds with evaluating the expression rhs and returns that result.

See also