Is an Argument a Filter Condition?

Is an Argument a Filter Condition?

Usage

assert_filter_cond(arg, optional = FALSE)

Arguments

arg

Quosure - filtering condition.

optional

Logical - is the argument optional? Defaults to FALSE.

Value

Performs necessary checks and returns arg if all pass. Otherwise throws an informative error.

Details

Check if arg is a suitable filtering condition to be used in functions like subset or dplyr::filter.

See also

Examples

library(admiral.test)
library(dplyr, warn.conflicts = FALSE)
library(rlang)
data(admiral_dm)

# typical usage in a function as a parameter check
example_fun <- function(dat, x) {
  x <- assert_filter_cond(enquo(x))
  filter(dat, !!x)
}

example_fun(admiral_dm, AGE == 64)
#> # A tibble: 5 x 25
#>   STUDYID   DOMAIN USUBJID   SUBJID RFSTDTC   RFENDTC  RFXSTDTC RFXENDTC RFICDTC
#>   <chr>     <chr>  <chr>     <chr>  <chr>     <chr>    <chr>    <chr>    <chr>  
#> 1 CDISCPIL… DM     01-701-1… 1023   2012-08-… 2012-09… 2012-08… 2012-09… ""     
#> 2 CDISCPIL… DM     01-703-1… 1042   2013-03-… 2013-08… 2013-03… 2013-08… ""     
#> 3 CDISCPIL… DM     01-706-1… 1041   2013-12-… 2014-07… 2013-12… 2014-07… ""     
#> 4 CDISCPIL… DM     01-713-1… 1179   2013-08-… 2014-02… 2013-08… 2014-02… ""     
#> 5 CDISCPIL… DM     01-717-1… 1344   2014-01-… 2014-03… 2014-01… 2014-03… ""     
#> # … with 16 more variables: RFPENDTC <chr>, DTHDTC <chr>, DTHFL <chr>,
#> #   SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>, ETHNIC <chr>,
#> #   ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>,
#> #   DMDTC <chr>, DMDY <dbl>

try(example_fun(admiral_dm, USUBJID))
#> Error in (function (e)  : Problem with `filter()` input `..1`.
#> ✖ Input `..1` must be a logical vector, not a character.
#> ℹ Input `..1` is `USUBJID`.