Expectation: Are Two Datasets Equal?

Uses diffdf::diffdf() to compares 2 datasets for any differences

expect_dfs_equal(base, compare, keys, ...)

Arguments

base

Input dataset

compare

Comparison dataset

keys

character vector of variables that define a unique row in the base and compare datasets

...

Additional arguments passed onto diffdf::diffdf()

Value

An error if base and compare do not match or NULL invisibly if they do

Author

Thomas Neitmann

Examples

if (FALSE) {
testthat::test_that("a missing row is detected", {
  data(dm)
  expect_dfs_equal(dm, dm[-1L, ], keys = "USUBJID")
})
}